moved components and conatiner folders to views directory
