+++ client/src/views/Profile.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import ProfileButton from '../components/ProfileButton';
-import ProfileDetails from '../containers/ProfileDetails';
+import ProfileButton from './components/ProfileButton';
+import ProfileDetails from './containers/ProfileDetails';
 import Logout from '../Logout';
-import GameList from '../components/GameList';
+import GameList from './components/GameList';
 
 const Profile = () => {
   const location = useLocation();
