+++ client/src/views/components/BehaviorInput.jsx
