+++ client/src/views/components/BehaviorInputContainer.jsx
