+++ client/src/views/components/GameList.jsx
