+++ client/src/views/containers/DebugPanel.jsx
